- Fix existing bugs

+++ src/main/java/backend/service/impl/TeacherServiceImpl.java
@@ -208,9 +208,10 @@ public class TeacherServiceImpl implements TeacherService {
                 if (teacherProfile.getAvatar() != null) {
                     String avatar = targetUser.getAvatar();
                     if (avatar != null && !avatar.isEmpty()) {
-                        CompletableFuture.runAsync(() ->
-                                fileManager.deleteFile(avatar)
-                        );
+                        if(!targetUser.getAvatar().equals("default.png"))
+                            CompletableFuture.runAsync(() ->
+                                    fileManager.deleteFile(avatar)
+                            );
                         targetUser.setAvatar(fileManager.uploadBase64Image(teacherProfile.getAvatar(), targetUser));
                     }
                 } else
